Two bug fixes + adding load_library
The first fix affects cases where a float has different numbers of
profiles available across profiles. Previously, the first profile
was always used to determine the names of the sensors. Now a profile
with the longest parameter string is used instead.

A second change is due to the fact that in some cases the index file
lists different modes than the corresponding Sprof file. The mode
matching will only be done based on the Sprof files for the time
being.

Check if presence of Rproj object in the Main_workshop before loading the functions

Add load_library function (removed form the Main_workshop)

Change the Setting verification in the load_float_data and in the show_ functions